backend/requirements.txt
openai>=1.30.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pyngrok==7.0.0
waitress==3.0.0
pyannote.audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0
pydub==0.25.1